oldlog add perms

+++ install.sh
@@ -2,8 +2,10 @@
 
 systemctl -q is-active log2zram  && { echo "ERROR: log2zram service is still running. Please run \"sudo service log2zram stop\" to stop it and uninstall"; exit 1; }
 [ "$(id -u)" -eq 0 ] || { echo "You need to be ROOT (sudo can be used)"; exit 1; }
+[ -d /usr/local/bin/log2zram ] && { echo "Log2Zram is already installed, uninstall first"; exit 1; }
 
-# log2ram
+
+# log2zram install 
 mkdir -p /usr/local/bin/log2zram
 install -m 644 log2zram.service /etc/systemd/system/log2zram.service
 install -m 755 log2zram /usr/local/bin/log2zram/log2zram
@@ -20,19 +22,21 @@ install -m 644 log2zram.logrotate /etc/logrotate.d/log2zram
 rm -rf /var/hdd.log
 mkdir -p /var/hdd.log
 mkdir -p /var/log/oldlog
-
-cp -rfp /var/log/*.1 /var/log/oldlog
-cp -rfp /var/log/*.gz /var/log/oldlog
-cp -rfp /var/log/*.old /var/log/oldlog
+chmod 754 /var/log/oldlog
+chown root:adm /var/log/oldlog
 # Prune logs
-rm -r /var/log/*.1
-rm -r /var/log/*.gz
-rm -r /var/log/*.old
+cp -au /var/log/*.1 /var/log/oldlog > /dev/null 2>&1 &
+cp -au /var/log/*.gz /var/log/oldlog > /dev/null 2>&1 &
+cp -au /var/log/*.old /var/log/oldlog > /dev/null 2>&1 &
+
+rm -r /var/log/*.1 > /dev/null 2>&1 &
+rm -r /var/log/*.gz > /dev/null 2>&1 &
+rm -r /var/log/*.old > /dev/null 2>&1 &
 # Clone /var/log
-# Prob better to use xcopy here with a --exclude
-rsync -arzh /var/log/ /var/hdd.log/
+echo "#####               Clone /var/log                 #####"
+rsync -arzhv /var/log/ /var/hdd.log/
 
 sed -i '/^include.*/i olddir /var/log/oldlog' /etc/logrotate.conf
 
-echo "#####          Reboot to activate log2ram         #####"
+echo "#####          Reboot to activate log2zram         #####"
 echo "##### edit /etc/log2zram.conf to configure options #####"
